Do not download assets in download operation
