+++ Sources/Safehill-Client/Tasks/Delegates/AssetDownloaderDelegate.swift
@@ -17,45 +17,10 @@ public protocol SHAssetDownloaderDelegate: SHInboundAssetOperationDelegate {
     func didReceiveRemoteAssetDescriptors(_ descriptors: [any SHAssetDescriptor],
                                           referencing users: [UserIdentifier: any SHServerUser])
     
-    /// Notifies about the start of a network download of a an asset from the CDN
-    /// - Parameter downloadRequest: the request
-    func didStartDownloadOfAsset(withGlobalIdentifier globalIdentifier: GlobalIdentifier,
-                                 descriptor: any SHAssetDescriptor,
-                                 in groupId: String)
-    
-    /// Notifies about an unrecoverable error (usually an asset that couldn't be downloaded after many attempts).
-    /// These assets global identifiers are blacklisted, to prevent the same failure to happen repeatedly.
-    /// It's the Client's responsibility to whitelist them as needed.
-    /// - Parameters:
-    ///   - globalIdentifier: the global identifier of the asset
-    ///   - groupId: the group id of the request it belongs to
-    func didBlacklistFailedToDownloadAsset(withGlobalIdentifier: GlobalIdentifier,
-                                          in groupId: String)
-    
     /// Notifies about assets in the local library that are linked to one on the server (backed up)
     /// - Parameters:
     ///   - localToGlobal: The global identifier of the remote asset to the corresponding local `PHAsset` from the Apple Photos Library
     func didIdentify(globalToLocalAssets: [GlobalIdentifier: PHAsset])
-    
-    /// Notifies about the successful download operation for a specific asset.
-    /// - Parameter decryptedAsset: the decrypted asset
-    /// - Parameter groupId: the group id of the request it belongs to
-    func didCompleteDownload(
-        of decryptedAsset: any SHDecryptedAsset,
-        in groupId: String
-    )
-    
-    /// One cycle of downloads has finished from local server
-    /// - Parameter localDescriptors: The descriptors for the assets downloaded from local server
-    func didCompleteDownloadCycle(
-        localAssetsAndDescriptors: [(any SHDecryptedAsset, any SHAssetDescriptor)]
-    )
-    
-    /// One cycle of downloads has finished from remote server
-    /// - Parameter localDescriptors: The descriptors for the assets downloaded from local server
-    func didCompleteDownloadCycle(
-        remoteAssetsAndDescriptors: [(any SHDecryptedAsset, any SHAssetDescriptor)]
-    )
     
     /// The download cycle failed
     /// - Parameter with: the error
